Step 1. Simple Read OK

+++ user/user.h
@@ -26,8 +26,8 @@ int getpid(void);
 char* sbrk(int);
 int sleep(int);
 int uptime(void);
-void *mmap(void *, uint64, int, int, int, uint64);
-int munmap(void *, uint64);
+void *mmap(void *, int, int, int, int, int);
+int munmap(void *, int);
 #ifdef LAB_NET
 int bind(uint16);
 int unbind(uint16);
